Code now utilizes a proxy server and accesses the server using Group Jade's API key.
RegisterActivity now stores user credentials on the remote server.
LoginActivity now queries the server for the inputted login data, and returns a token if valid.
The majority of my code is now commented if clarification is needed.
API Key is saved in strings.xml and can be accessed via the id: API_KEY